Move tests for distinct values to a separate file for better organization.
