+++ metricflow/test/plan_conversion/test_dataflow_to_sql_plan.py
@@ -1068,29 +1068,3 @@ def test_combine_output_node(  # noqa: D
         sql_client=sql_client,
         node=combine_output_node,
     )
-
-
-@pytest.mark.sql_engine_snapshot
-def test_dimensions_requiring_join(
-    request: FixtureRequest,
-    mf_test_session_state: MetricFlowTestSessionState,
-    dataflow_plan_builder: DataflowPlanBuilder,
-    dataflow_to_sql_converter: DataflowToSqlQueryPlanConverter,
-    sql_client: SqlClient,
-) -> None:
-    """Tests querying 2 dimensions that require a join."""
-    dimension_specs = (
-        DimensionSpec(element_name="home_state_latest", entity_links=(EntityReference(element_name="user"),)),
-        DimensionSpec(element_name="is_lux_latest", entity_links=(EntityReference(element_name="listing"),)),
-    )
-    dataflow_plan = dataflow_plan_builder.build_plan_for_distinct_values(
-        query_spec=MetricFlowQuerySpec(dimension_specs=dimension_specs)
-    )
-
-    convert_and_check(
-        request=request,
-        mf_test_session_state=mf_test_session_state,
-        dataflow_to_sql_converter=dataflow_to_sql_converter,
-        sql_client=sql_client,
-        node=dataflow_plan.sink_output_nodes[0].parent_node,
-    )
